move visit_0xfd_operator into separate file
We do this to further improve compile times for when the `simd` crate feature is disabled.
